qusai_whisper_experiment/requirements.txt
openai-whisper>=20240930
torch>=2.0.0
numpy>=1.20.0
tqdm>=4.65.0
pathlib>=1.0.0
argparse>=1.4.0
pydub>=0.25.0
soundfile>=0.12.0 